+++ ButtonCore/src/main/java/buttondevteam/component/channel/ChannelComponent.java
@@ -1,0 +1,26 @@
+package buttondevteam.component.channel;
+
+import buttondevteam.lib.architecture.Component;
+import org.bukkit.plugin.java.JavaPlugin;
+
+public class ChannelComponent extends Component {
+	@Override
+	protected void register(JavaPlugin plugin) {
+		super.register(plugin);
+	}
+
+	@Override
+	protected void unregister(JavaPlugin plugin) {
+		super.unregister(plugin);
+	}
+
+	@Override
+	protected void enable() {
+
+	}
+
+	@Override
+	protected void disable() {
+
+	}
+}
