+++ ButtonCore/src/main/java/buttondevteam/component/channel/ChatChannelRegisterEvent.java
@@ -1,4 +1,4 @@
-package buttondevteam.lib.chat;
+package buttondevteam.component.channel;
 
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
