--- ButtonCore/src/main/java/buttondevteam/component/commands/CommandComponent.java
@@ -1,15 +1,0 @@
-package buttondevteam.component.commands;
-
-import buttondevteam.lib.architecture.Component;
-
-public class CommandComponent extends Component { //TODO: Do we just move everything here?
-	@Override
-	public void enable() {
-
-	}
-
-	@Override
-	public void disable() {
-
-	}
-}
