+++ ButtonCore/src/main/java/buttondevteam/component/updater/PluginUpdaterComponent.java
@@ -6,11 +6,11 @@ import buttondevteam.lib.chat.TBMCChatAPI;
 public class PluginUpdaterComponent extends Component {
 	@Override
 	public void enable() {
-		TBMCChatAPI.AddCommand(getPlugin(), UpdatePluginCommand.class);
+		TBMCChatAPI.AddCommand(this, new UpdatePluginCommand());
 	}
 
 	@Override
-	public void disable() { //TODO: Unregister commands and such
+	public void disable() { //Commands are automatically unregistered
 
 	}
 }
