+++ ButtonCore/src/main/java/buttondevteam/lib/TBMCSystemChatEvent.java
@@ -1,6 +1,6 @@
 package buttondevteam.lib;
 
-import buttondevteam.lib.chat.Channel;
+import buttondevteam.component.channel.Channel;
 import lombok.Getter;
 import org.bukkit.command.CommandSender;
 import org.bukkit.event.HandlerList;
